requirements.txt
fastapi==0.94.0
Pillow~=9.4.0
numpy~=1.24.2
opencv-python~=4.7.0.72